refactor: move out api config
